move JSPath and Proxy to Config
