+++ src/PostgREST/OpenAPI.hs
@@ -22,7 +22,7 @@ import Control.Lens (at, (.~), (?~))
 import Data.Swagger
 
 import PostgREST.Config               (AppConfig (..), Proxy (..),
-                                       docsVersion, prettyVersion)
+                                       docsVersion, prettyVersion, isMalformedProxyUri, toURI)
 import PostgREST.DbStructure          (DbStructure (..), tableCols,
                                        tablePKCols)
 import PostgREST.DbStructure.Proc     (PgArg (..),
@@ -30,7 +30,6 @@ import PostgREST.DbStructure.Proc     (PgArg (..),
 import PostgREST.DbStructure.Relation (PrimaryKey (..))
 import PostgREST.DbStructure.Table    (Column (..), ForeignKey (..),
                                        Table (..))
-import PostgREST.ProxyUri             (isMalformedProxyUri, toURI)
 
 import PostgREST.ContentType
